feat: [M3-6967] consolidate tests and utils
