+++ packages/manager/src/features/Backups/utils.test.ts
@@ -1,6 +1,4 @@
-import { linodeFactory, linodeTypeFactory } from 'src/factories';
-
-import { getFailureNotificationText, getTotalBackupsPrice } from './utils';
+import { getFailureNotificationText } from './utils';
 
 describe('getFailureNotificationText', () => {
   it('has message for when all enables fail', () => {
@@ -24,56 +22,5 @@ describe('getFailureNotificationText', () => {
     expect(
       getFailureNotificationText({ failedCount: 1, successCount: 5 })
     ).toBe('Enabled backups successfully for 5 Linodes, but 1 Linode failed.');
-  });
-});
-
-describe('getTotalBackupsPrice', () => {
-  it('correctly calculates the total price for Linode backups', () => {
-    const linodes = linodeFactory.buildList(3, { type: 'my-type' });
-    const types = linodeTypeFactory.buildList(1, {
-      addons: { backups: { price: { monthly: 2.5 } } },
-      id: 'my-type',
-    });
-    expect(
-      getTotalBackupsPrice({
-        flags: { dcSpecificPricing: false },
-        linodes,
-        types,
-      })
-    ).toBe(7.5);
-  });
-
-  it('correctly calculates the total price with DC-specific pricing for Linode backups', () => {
-    const basePriceLinodes = linodeFactory.buildList(2, { type: 'my-type' });
-    const priceIncreaseLinode = linodeFactory.build({
-      region: 'id-cgk',
-      type: 'my-type',
-    });
-    const linodes = [...basePriceLinodes, priceIncreaseLinode];
-    const types = linodeTypeFactory.buildList(1, {
-      addons: {
-        backups: {
-          price: {
-            hourly: 0.004,
-            monthly: 2.5,
-          },
-          region_prices: [
-            {
-              hourly: 0.0048,
-              id: 'id-cgk',
-              monthly: 3.57,
-            },
-          ],
-        },
-      },
-      id: 'my-type',
-    });
-    expect(
-      getTotalBackupsPrice({
-        flags: { dcSpecificPricing: true },
-        linodes,
-        types,
-      })
-    ).toBe(8.57);
   });
 });
