+++ packages/manager/src/utilities/pricing/backups.ts
@@ -1,9 +1,32 @@
-import { LinodeType, PriceObject } from '@linode/api-v4';
+import type { ExtendedType } from '../extendType';
+import type { Linode, LinodeType, PriceObject } from '@linode/api-v4';
+import type { FlagSet } from 'src/featureFlags';
 
-import { FlagSet } from 'src/featureFlags';
+/**
+ * Gets the backup price of a Linode type for a specific region.
+ *
+ * @param type The Linode Type
+ * @param regionId The region to get the price for
+ * @returns backup pricing information for this specific linode type in a region
+ */
+export const getLinodeBackupPrice = (
+  type: LinodeType,
+  regionId: string
+): PriceObject => {
+  const regionSpecificBackupPrice = type.addons.backups.region_prices?.find(
+    (regionPrice) => regionPrice.id === regionId
+  );
 
-import { ExtendedType } from '../extendType';
-import { getLinodeBackupPrice } from './linodes';
+  if (regionSpecificBackupPrice) {
+    return {
+      hourly: regionSpecificBackupPrice.hourly,
+      monthly: regionSpecificBackupPrice.monthly,
+    };
+  }
+
+  // TODO: M3-7063 (defaults)
+  return type.addons.backups.price;
+};
 
 interface BackupsPriceOptions {
   flags: FlagSet;
@@ -20,10 +43,46 @@ export const getMonthlyBackupsPrice = ({
   type,
 }: BackupsPriceOptions): PriceObject['monthly'] => {
   if (!region || !type) {
+    // TODO: M3-7063 (defaults)
     return 0;
   }
 
   return flags.dcSpecificPricing
     ? getLinodeBackupPrice(type, region).monthly
     : type?.addons.backups.price.monthly;
+};
+
+export interface TotalBackupsPriceOptions {
+  /**
+   * Our feature flags so we can determined whether or not to add price increase.
+   * @example { dcSpecificPricing: true }
+   */
+  flags: FlagSet;
+  /**
+   * List of linodes without backups enabled
+   */
+  linodes: Linode[];
+  /**
+   * List of types for the linodes without backups
+   */
+  types: LinodeType[];
+}
+
+export const getTotalBackupsPrice = ({
+  flags,
+  linodes,
+  types,
+}: TotalBackupsPriceOptions) => {
+  return linodes.reduce((prevValue: number, linode: Linode) => {
+    const type = types.find((type) => type.id === linode.type);
+
+    const backupsMonthlyPrice: PriceObject['monthly'] =
+      getMonthlyBackupsPrice({
+        flags,
+        region: linode.region,
+        type,
+      }) || 0;
+
+    return prevValue + backupsMonthlyPrice;
+  }, 0);
 };
