+++ packages/manager/src/utilities/pricing/linodes.ts
@@ -24,32 +24,8 @@ export const getLinodeRegionPrice = (
     };
   }
 
+  // TODO: M3-7063 (defaults)
   return type.price;
-};
-
-/**
- * Gets the backup price of a Linode type for a specific region.
- *
- * @param type The Linode Type
- * @param regionId The region to get the price for
- * @returns backup pricing information for this specific linode type in a region
- */
-export const getLinodeBackupPrice = (
-  type: LinodeType,
-  regionId: string
-): PriceObject => {
-  const regionSpecificBackupPrice = type.addons.backups.region_prices?.find(
-    (regionPrice) => regionPrice.id === regionId
-  );
-
-  if (regionSpecificBackupPrice) {
-    return {
-      hourly: regionSpecificBackupPrice.hourly,
-      monthly: regionSpecificBackupPrice.monthly,
-    };
-  }
-
-  return type.addons.backups.price;
 };
 
 interface IsPriceDifferentOptions {
